Login Signup Form
Fully Functional Login Signup form. When a new account is created the data stores in a txt file. When user try to login it matches the username and password with the existing accounts in the file if it matches then login become successful otherwise it throws and error. Validation is also used in the form like email validation and password must at least 8 characters long. It also has the option to reset password that asks user to add username and new password if username matches with any existing account then the new password got updated with the previous one and user can login with the latest password"